refactoring project structure.
This project now includes a web application and a file service. That means that it is now a multi-module project.

the file-service and app working together.

And i removed all security-protection and keycloak.

I have added 2 profiles:
The first one is "dev" for working with docker-compose and dev-tools.

The second one is "prod" to work with dockerfile + docker-compose in the parent project to create one container for the entire rental service.

coming soon:
tests, new security, fixes and refactoring with classes, DTOs, payload